# Cache the derivatives of the barycentres