Before:
```
2023-06-08T20:52:51+02:00
Running C:\Users\robin\Projects\mockingbirdnest\Principia\Release\x64\benchmarks.exe
Run on (12 X 2688 MHz CPU s)
CPU Caches:
  L1 Data 48 KiB (x6)
  L1 Instruction 32 KiB (x6)
  L2 Unified 1280 KiB (x6)
  L3 Unified 12288 KiB (x1)
-------------------------------------------------------------------------------------
Benchmark                                           Time             CPU   Iterations
-------------------------------------------------------------------------------------
LagrangeEquipotentialsBenchmark/EarthMoon   399438500 ns    328125000 ns            2
LagrangeEquipotentialsBenchmark/SunNeptune 8773415400 ns   5171875000 ns            1
```

After:
```
2023-06-08T20:50:40+02:00
Running C:\Users\robin\Projects\mockingbirdnest\Principia\Release\x64\benchmarks.exe
Run on (12 X 2688 MHz CPU s)
CPU Caches:
  L1 Data 48 KiB (x6)
  L1 Instruction 32 KiB (x6)
  L2 Unified 1280 KiB (x6)
  L3 Unified 12288 KiB (x1)
-------------------------------------------------------------------------------------
Benchmark                                           Time             CPU   Iterations
-------------------------------------------------------------------------------------
LagrangeEquipotentialsBenchmark/EarthMoon    68671460 ns     50000000 ns           10
LagrangeEquipotentialsBenchmark/SunNeptune   82250800 ns     56250000 ns           10
```

399438500/68671460 = about 6 times faster for EarthMoo,
8773415400/82250800 = about 100 times faster for SunNeptune.